requirements.txt
percept
praw==2.1.3
Scrapy==0.16.5
ipython==0.13.2
nltk==2.0.4
pandas==0.11.0
scikit-learn==0.13.1
scipy==0.12.0
fisher==0.1.4
scikits.audiolab
